changed Problem and adjusted the project for it

+++ include/Solution.h
@@ -10,7 +10,7 @@ class Solution
         virtual ~Solution();
         int* Genome;
         int size;
-        int eval;
+        double eval;
         print();
 
     protected:
